requirements.txt
pandas==2.3.0
python-dotenv==1.1.1
